+++ CS2D-internal/mem.h
@@ -9,5 +9,5 @@ namespace mem
 	void PatchEx(BYTE* dst, BYTE* src, unsigned int size, HANDLE hProcess);
 	void Nop(BYTE* dst, unsigned int size);
 	void NopEx(BYTE* dst, unsigned int size, HANDLE hProcess);
-	uintptr_t FindDMAAddy(uintptr_t ptr, std::vector<unsigned int> offsets);
+	uintptr_t FindDMAAddy(uintptr_t ptr, std::vector<uintptr_t> offsets);
 }
